requirements.txt
aiohttp
requests
python-whois
dnspython
